// frontend/src/app/models/item-value.ts
export interface ItemValue {
    name: string;
    detail: string;
    type: string;
    weight: number;
    value: number;
    properties: string[];
    dmg1: string;
    dmg2: string;
    dmgType: string;
    range: string;
    ac: number;
    stealth: boolean;
    magic: boolean;
    strength: number;
    text: string;
}